♻️ Move custom payment method config to its own /config/payment page
